refactor(java): Move examples to new project to depend on ESDK (#1441)
