+++ AwsCryptographicMaterialProviders/runtimes/java/build.gradle.kts
@@ -1,3 +1,5 @@
+// Copyright Amazon.com Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: CC-BY-SA-4.0
 import java.io.File
 import java.io.FileInputStream
 import java.util.Properties
@@ -22,29 +24,6 @@ version = props.getProperty("mplVersion")
 description = "AWS Cryptographic Material Providers Library"
 var slf4jVersion = "1.7.32"
 
-sourceSets {
-    create("examples") {
-        compileClasspath += sourceSets.main.get().output
-        runtimeClasspath += sourceSets.main.get().output
-    }
-    create("testExamples") {
-        compileClasspath += sourceSets.test.get().output + sourceSets["examples"].output + sourceSets.main.get().output
-        runtimeClasspath += sourceSets.test.get().output + sourceSets["examples"].output + sourceSets.main.get().output
-    }
-}
-val examplesImplementation: Configuration by configurations.getting{
-    extendsFrom(configurations.testImplementation.get())
-}
-configurations.add(examplesImplementation)
-val examplesAnnotationProcessor: Configuration by configurations.getting{
-    extendsFrom(configurations.testAnnotationProcessor.get())
-}
-configurations.add(examplesAnnotationProcessor)
-val testExamplesImplementation: Configuration by configurations.getting{
-    extendsFrom(configurations["examplesImplementation"])
-}
-configurations.add(testExamplesImplementation)
-
 java {
     toolchain.languageVersion.set(JavaLanguageVersion.of(8))
     sourceSets["main"].java {
@@ -53,43 +32,13 @@ java {
     sourceSets["test"].java {
         srcDir("src/test")
     }
-    sourceSets["examples"].java {
-        srcDir("src/examples")
-    }
-    sourceSets["testExamples"].java {
-        srcDir("src/testExamples")
-    }
     withJavadocJar()
     withSourcesJar()
-}
-
-var caUrl: URI? = null
-@Nullable
-val caUrlStr: String? = System.getenv("CODEARTIFACT_URL_JAVA_CONVERSION")
-if (!caUrlStr.isNullOrBlank()) {
-    caUrl = URI.create(caUrlStr)
-}
-
-var caPassword: String? = null
-@Nullable
-val caPasswordString: String? = System.getenv("CODEARTIFACT_AUTH_TOKEN")
-if (!caPasswordString.isNullOrBlank()) {
-    caPassword = caPasswordString
 }
 
 repositories {
     mavenCentral()
     mavenLocal()
-    if (caUrl != null && caPassword != null) {
-        maven {
-            name = "CodeArtifact"
-            url = caUrl!!
-            credentials {
-                username = "aws"
-                password = caPassword!!
-            }
-        }
-    }
 }
 
 dependencies {
@@ -118,16 +67,6 @@ dependencies {
     testImplementation("org.slf4j:slf4j-api:${slf4jVersion}")
     testImplementation("org.slf4j:slf4j-simple:${slf4jVersion}")
     testImplementation("org.slf4j:jcl-over-slf4j:${slf4jVersion}")
-        
-    // Example Dependencies
-    examplesImplementation("software.amazon.awssdk:arns")
-    examplesImplementation("software.amazon.awssdk:auth")
-    examplesImplementation("software.amazon.awssdk:sts")
-    examplesImplementation("software.amazon.awssdk:utils")
-    examplesImplementation("software.amazon.awssdk:apache-client")
-    examplesAnnotationProcessor("org.projectlombok:lombok:1.18.30")
-    examplesImplementation("com.google.code.findbugs:jsr305:3.0.2")
-
 }
 
 publishing {
@@ -299,46 +238,6 @@ tasks.test {
             }
         }
     })
-}
-
-val testExamples = task<Test>("testExamples") {
-    description = "Runs examples tests."
-    group = "verification"
-
-    testClassesDirs = sourceSets["testExamples"].output.classesDirs
-    classpath = sourceSets["testExamples"].runtimeClasspath + sourceSets["examples"].output + sourceSets.main.get().output
-    shouldRunAfter("compileJava", "compileExamplesJava", "test")
-    // This will show System.out.println statements
-    testLogging.showStandardStreams = true
-    useTestNG()
-
-    testLogging {
-        events("passed")
-    }
-    filter {
-        excludeTestsMatching("software.amazon.cryptography.example.hierarchy.concurrent.*")
-    }
-}
-
-val testConcurrentExamples = task<Test>("testConcurrentExamples") {
-    description = "Runs concurrency tests."
-    group = "verification"
-
-    testClassesDirs = sourceSets["testExamples"].output.classesDirs
-    classpath = sourceSets["testExamples"].runtimeClasspath + sourceSets["examples"].output + sourceSets.main.get().output
-    // This will show System.out.println statements
-    testLogging.showStandardStreams = true
-    useTestNG() {
-        suites("src/testExamples/java/software/amazon/cryptography/example/hierarchy/concurrent/testng-parallel.xml")
-        maxParallelForks = 2
-    }
-
-    testLogging {
-        events("passed")
-    }
-    filter {
-        includeTestsMatching("software.amazon.cryptography.example.hierarchy.concurrent.*")
-    }
 }
 
 fun buildPom(mavenPublication: MavenPublication) {
