+++ Examples/java/.gitignore
@@ -1,0 +1,12 @@
+# Ignore Gradle project-specific cache directory
+.gradle
+
+# Ignore Gradle build output directory
+build
+
+# JetBrains
+.idea/*
+*.iml
+
+# Mac OS X
+.DS_Store
