+++ Examples/java/settings.gradle.kts
@@ -1,0 +1,3 @@
+// Copyright Amazon.com Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: CC-BY-SA-4.0
+rootProject.name = "AwsCryptographicMaterialProviderLibraryExamples"
