+++ Examples/java/src/main/java/software/amazon/cryptography/example/Constants.java
