+++ Examples/java/src/main/java/software/amazon/cryptography/example/hierarchy/KeyStoreProvider.java
