+++ Examples/java/src/main/java/software/amazon/cryptography/example/hierarchy/mutations/MutationDecryptEncryptExample.java
