+++ Examples/java/src/test/java/software/amazon/cryptography/example/hierarchy/ValidateKeyStoreItem.java
